chore: refactor to db dump command
