+++ internal/db/dump/dump.go
@@ -1,4 +1,4 @@
-package pull
+package dump
 
 import (
 	"context"
@@ -6,19 +6,17 @@ import (
 	"errors"
 	"fmt"
 	"os"
-	"path/filepath"
 	"strings"
 
 	"github.com/spf13/afero"
-	"github.com/supabase/cli/internal/migration/new"
 	"github.com/supabase/cli/internal/utils"
 )
 
 //go:embed templates/dump_initial_migration.sh
 var dumpInitialMigrationScript string
 
-func Run(ctx context.Context, username, password, database, host string, fsys afero.Fs) error {
-	fmt.Fprintln(os.Stderr, "Pulling schemas from remote database...")
+func Run(ctx context.Context, path, username, password, database, host string, fsys afero.Fs) error {
+	fmt.Fprintln(os.Stderr, "Dumping schemas from remote database...")
 	out, err := utils.DockerRunOnce(ctx, utils.Pg14Image, []string{
 		"PGHOST=" + host,
 		"PGUSER=" + username,
@@ -30,15 +28,14 @@ func Run(ctx context.Context, username, password, database, host string, fsys af
 		return errors.New("Error running pg_dump on remote database: " + err.Error())
 	}
 
-	path := new.GetMigrationPath("remote_commit")
-	if err := utils.MkdirIfNotExistFS(fsys, filepath.Dir(path)); err != nil {
-		return err
-	}
-	if err := afero.WriteFile(fsys, path, []byte(out), 0644); err != nil {
-		return err
+	if len(path) > 0 {
+		if err := afero.WriteFile(fsys, path, []byte(out), 0644); err != nil {
+			return err
+		}
+		fmt.Fprintln(os.Stderr, "Dumped schema to "+utils.Bold(path)+".")
+	} else {
+		fmt.Println(out)
 	}
 
-	fmt.Fprintln(os.Stderr, "Created new migration at "+utils.Bold(path)+".")
-	fmt.Println("Finished " + utils.Aqua("supabase migration pull") + ".")
 	return nil
 }
