+++ internal/db/dump/templates/dump_initial_migration.sh
